fix: reduce rerenders
